requirements.txt
streamlit==1.44.1
pandas==2.2.3
python-dotenv==1.0.1
requests==2.32.3
reportlab==4.2.5
python-docx==1.1.2
beautifulsoup4==4.12.3
